renamed services to user-service, made index.js at user-service
