+++ __test__/user-endpoint.spec.js
@@ -1,6 +1,6 @@
 const makeFakeUser = require('./fixtures/user')
 const makeUserDb = require('./stubs/userdb-stub')
-const makeUserActions = require('../src/domain/services/user-actions')
+const makeUserActions = require('../src/domain/user-service/user-actions')
 const makeUsersEndpointHandler = require('../src/controllers/user-endpoint')
