+++ src/domain/user-service/index.js
@@ -1,0 +1,6 @@
+const makeUserActions = require('./user-actions');
+const userDb = require('../../db/userdb');
+
+const usersService = makeUserActions({ userDb });
+
+module.exports = usersService;
